refactor(server): modularize routes and ws bridge architecture
