+++ web/server/routes/fs-routes.ts
@@ -1,0 +1,281 @@
+import { execSync } from "node:child_process";
+import { readdir, readFile, stat, writeFile, mkdir } from "node:fs/promises";
+import { homedir } from "node:os";
+import { dirname, join, resolve } from "node:path";
+import type { Hono } from "hono";
+
+function execCaptureStdout(
+  command: string,
+  options: { cwd: string; encoding: "utf-8"; timeout: number },
+): string {
+  try {
+    return execSync(command, options);
+  } catch (err: unknown) {
+    const maybe = err as { stdout?: Buffer | string };
+    if (typeof maybe.stdout === "string") return maybe.stdout;
+    if (maybe.stdout && Buffer.isBuffer(maybe.stdout)) {
+      return maybe.stdout.toString("utf-8");
+    }
+    throw err;
+  }
+}
+
+function resolveBranchDiffBases(repoRoot: string): string[] {
+  const options = { cwd: repoRoot, encoding: "utf-8", timeout: 5000 } as const;
+
+  try {
+    const originHead = execSync("git symbolic-ref refs/remotes/origin/HEAD", options).trim();
+    const match = originHead.match(/^refs\/remotes\/origin\/(.+)$/);
+    if (match?.[1]) {
+      return [`origin/${match[1]}`, match[1]];
+    }
+  } catch {
+    // No remote HEAD ref available, fallback to common local defaults.
+  }
+
+  try {
+    const branches = execSync("git branch --list main master", options).trim();
+    if (branches.includes("main")) return ["main"];
+    if (branches.includes("master")) return ["master"];
+  } catch {
+    // Ignore and use a conservative fallback below.
+  }
+
+  return ["main"];
+}
+
+export function registerFsRoutes(api: Hono): void {
+  api.get("/fs/list", async (c) => {
+    const rawPath = c.req.query("path") || homedir();
+    const basePath = resolve(rawPath);
+    try {
+      const entries = await readdir(basePath, { withFileTypes: true });
+      const dirs: { name: string; path: string }[] = [];
+      for (const entry of entries) {
+        if (entry.isDirectory() && !entry.name.startsWith(".")) {
+          dirs.push({ name: entry.name, path: join(basePath, entry.name) });
+        }
+      }
+      dirs.sort((a, b) => a.name.localeCompare(b.name));
+      return c.json({ path: basePath, dirs, home: homedir() });
+    } catch {
+      return c.json(
+        {
+          error: "Cannot read directory",
+          path: basePath,
+          dirs: [],
+          home: homedir(),
+        },
+        400,
+      );
+    }
+  });
+
+  api.get("/fs/home", (c) => {
+    const home = homedir();
+    const cwd = process.cwd();
+    // Only report cwd if the user launched companion from a real project directory
+    // (not from the package root or the home directory itself)
+    const packageRoot = process.env.__COMPANION_PACKAGE_ROOT;
+    const isProjectDir =
+      cwd !== home &&
+      (!packageRoot || !cwd.startsWith(packageRoot));
+    return c.json({ home, cwd: isProjectDir ? cwd : home });
+  });
+
+  api.get("/fs/tree", async (c) => {
+    const rawPath = c.req.query("path");
+    if (!rawPath) return c.json({ error: "path required" }, 400);
+    const basePath = resolve(rawPath);
+
+    interface TreeNode {
+      name: string;
+      path: string;
+      type: "file" | "directory";
+      children?: TreeNode[];
+    }
+
+    async function buildTree(dir: string, depth: number): Promise<TreeNode[]> {
+      if (depth > 10) return [];
+      try {
+        const entries = await readdir(dir, { withFileTypes: true });
+        const nodes: TreeNode[] = [];
+        for (const entry of entries) {
+          if (entry.name.startsWith(".") || entry.name === "node_modules") continue;
+          const fullPath = join(dir, entry.name);
+          if (entry.isDirectory()) {
+            const children = await buildTree(fullPath, depth + 1);
+            nodes.push({
+              name: entry.name,
+              path: fullPath,
+              type: "directory",
+              children,
+            });
+          } else if (entry.isFile()) {
+            nodes.push({ name: entry.name, path: fullPath, type: "file" });
+          }
+        }
+        nodes.sort((a, b) => {
+          if (a.type !== b.type) return a.type === "directory" ? -1 : 1;
+          return a.name.localeCompare(b.name);
+        });
+        return nodes;
+      } catch {
+        return [];
+      }
+    }
+
+    const tree = await buildTree(basePath, 0);
+    return c.json({ path: basePath, tree });
+  });
+
+  api.get("/fs/read", async (c) => {
+    const filePath = c.req.query("path");
+    if (!filePath) return c.json({ error: "path required" }, 400);
+    const absPath = resolve(filePath);
+    try {
+      const info = await stat(absPath);
+      if (info.size > 2 * 1024 * 1024) {
+        return c.json({ error: "File too large (>2MB)" }, 413);
+      }
+      const content = await readFile(absPath, "utf-8");
+      return c.json({ path: absPath, content });
+    } catch (e: unknown) {
+      return c.json(
+        { error: e instanceof Error ? e.message : "Cannot read file" },
+        404,
+      );
+    }
+  });
+
+  api.put("/fs/write", async (c) => {
+    const body = await c.req.json().catch(() => ({}));
+    const { path: filePath, content } = body;
+    if (!filePath || typeof content !== "string") {
+      return c.json({ error: "path and content required" }, 400);
+    }
+    const absPath = resolve(filePath);
+    try {
+      await writeFile(absPath, content, "utf-8");
+      return c.json({ ok: true, path: absPath });
+    } catch (e: unknown) {
+      return c.json(
+        { error: e instanceof Error ? e.message : "Cannot write file" },
+        500,
+      );
+    }
+  });
+
+  api.get("/fs/diff", (c) => {
+    const filePath = c.req.query("path");
+    if (!filePath) return c.json({ error: "path required" }, 400);
+    const base = c.req.query("base");
+    const absPath = resolve(filePath);
+    try {
+      const repoRoot = execSync("git rev-parse --show-toplevel", {
+        cwd: dirname(absPath),
+        encoding: "utf-8",
+        timeout: 5000,
+      }).trim();
+      const relPath = execSync(`git -C "${repoRoot}" ls-files --full-name -- "${absPath}"`, {
+        encoding: "utf-8",
+        timeout: 5000,
+      }).trim() || absPath;
+
+      let diff = "";
+
+      if (base === "default-branch") {
+        const diffBases = resolveBranchDiffBases(repoRoot);
+        for (const b of diffBases) {
+          try {
+            diff = execCaptureStdout(`git diff ${b} -- "${relPath}"`, {
+              cwd: repoRoot,
+              encoding: "utf-8",
+              timeout: 5000,
+            });
+            break;
+          } catch {
+            // If a base ref is unavailable, try the next candidate.
+          }
+        }
+      } else {
+        try {
+          diff = execCaptureStdout(`git diff HEAD -- "${relPath}"`, {
+            cwd: repoRoot,
+            encoding: "utf-8",
+            timeout: 5000,
+          });
+        } catch {
+          // HEAD may not exist in a fresh repo with no commits; fall through to untracked handling.
+        }
+      }
+
+      if (!diff.trim()) {
+        const untracked = execSync(`git ls-files --others --exclude-standard -- "${relPath}"`, {
+          cwd: repoRoot,
+          encoding: "utf-8",
+          timeout: 5000,
+        }).trim();
+        if (untracked) {
+          diff = execCaptureStdout(`git diff --no-index -- /dev/null "${absPath}"`, {
+            cwd: repoRoot,
+            encoding: "utf-8",
+            timeout: 5000,
+          });
+        }
+      }
+
+      return c.json({ path: absPath, diff });
+    } catch {
+      return c.json({ path: absPath, diff: "" });
+    }
+  });
+
+  api.get("/fs/claude-md", async (c) => {
+    const cwd = c.req.query("cwd");
+    if (!cwd) return c.json({ error: "cwd required" }, 400);
+    const resolvedCwd = resolve(cwd);
+    const candidates = [
+      join(resolvedCwd, "CLAUDE.md"),
+      join(resolvedCwd, ".claude", "CLAUDE.md"),
+    ];
+
+    const files: { path: string; content: string }[] = [];
+    for (const p of candidates) {
+      try {
+        const content = await readFile(p, "utf-8");
+        files.push({ path: p, content });
+      } catch {
+        // file doesn't exist — skip
+      }
+    }
+
+    return c.json({ cwd: resolvedCwd, files });
+  });
+
+  api.put("/fs/claude-md", async (c) => {
+    const body = await c.req.json().catch(() => ({}));
+    const { path: filePath, content } = body;
+    if (!filePath || typeof content !== "string") {
+      return c.json({ error: "path and content required" }, 400);
+    }
+    const base = filePath.split("/").pop();
+    if (base !== "CLAUDE.md") {
+      return c.json({ error: "Can only write CLAUDE.md files" }, 400);
+    }
+    const absPath = resolve(filePath);
+    if (!absPath.endsWith("/CLAUDE.md") && !absPath.endsWith("/.claude/CLAUDE.md")) {
+      return c.json({ error: "Invalid CLAUDE.md path" }, 400);
+    }
+    try {
+      await mkdir(dirname(absPath), { recursive: true });
+      await writeFile(absPath, content, "utf-8");
+      return c.json({ ok: true, path: absPath });
+    } catch (e: unknown) {
+      return c.json(
+        { error: e instanceof Error ? e.message : "Cannot write file" },
+        500,
+      );
+    }
+  });
+}
